+++ packages/terra-form/src/Select.jsx
@@ -1,9 +1,11 @@
 import React from 'react';
 import PropTypes from 'prop-types';
-import classNames from 'classnames';
-import 'terra-base/lib/baseStyles';
+import classNames from 'classnames/bind';
 
-import './Select.scss';
+import 'terra-base/lib/baseStyles';
+import styles from './Select.scss';
+
+const cx = classNames.bind(styles);
 
 const propTypes = {
 
@@ -69,10 +71,11 @@ const Select = ({
   value,
   ...customProps
 }) => {
-  const additionalSelectFieldProps = Object.assign({}, customProps);
+  const additionalSelectProps = Object.assign({}, customProps);
+  const selectClasses = cx(['select', additionalSelectProps.className]);
 
   if (required) {
-    additionalSelectFieldProps['aria-required'] = 'true';
+    additionalSelectProps['aria-required'] = 'true';
   }
 
   if (choices && options === undefined) {
@@ -85,13 +88,13 @@ const Select = ({
 
   return (
     <select
-      {...additionalSelectFieldProps}
+      {...additionalSelectProps}
       name={name}
       required={required}
       onChange={onChange}
       defaultValue={defaultValue}
       value={value}
-      className={classNames('terra-Form-select', additionalSelectFieldProps.className)}
+      className={selectClasses}
     >
       {options.map(option => <option key={`${option.value}-${option.display}`} value={option.value}>{option.display}</option>)}
     </select>