+++ packages/terra-form/tests/nightwatch/textarea-field/DefaultTextareaField.jsx
@@ -2,6 +2,6 @@ import React from 'react';
 
 import TextareaField from '../../../lib/TextareaField';
 
-const textareaField = () => <TextareaField />;
+const textareaField = () => <TextareaField id="default-textarea-field" />;
 
 export default textareaField;