+++ packages/terra-image/tests/jest/Image.test.jsx
@@ -12,14 +12,12 @@ const image = <Image src="" alt="circle" variant="circle" isFluid />;
 /* Default component */
 it('should render a default component', () => {
   const wrapper = shallow(<Image src="" />);
-  expect(wrapper.prop('className')).toContain('terra-Image--default');
   expect(wrapper.prop('alt')).toEqual(' ');
 });
 
 /* Default component with a height and width */
 it('should render a default image component with a specified height and width', () => {
   const wrapper = shallow(<Image src="" height={'200'} width={'300'} alt="default image" />);
-  expect(wrapper.prop('className')).toContain('terra-Image--default');
   expect(wrapper.prop('alt')).toEqual('default image');
   expect(wrapper.prop('height')).toEqual('200');
   expect(wrapper.prop('width')).toEqual('300');
@@ -28,34 +26,31 @@ it('should render a default image component with a specified height and width', 
 /* Image with a rounded variant */
 it('should render a rounded image component', () => {
   const wrapper = shallow(<Image src="" variant="rounded" alt="rounded image" />);
-  expect(wrapper.prop('className')).toContain('terra-Image--rounded');
   expect(wrapper.prop('alt')).toEqual('rounded image');
 });
 
 /* Image with a circle variant */
 it('should render a circle image component', () => {
   const wrapper = shallow(<Image src="" variant="circle" alt="circle image" />);
-  expect(wrapper.prop('className')).toContain('terra-Image--circle');
   expect(wrapper.prop('alt')).toEqual('circle image');
 });
 
 /* Image with a thumbnail variant */
 it('should render a thumbnail image component', () => {
   const wrapper = shallow(<Image src="" variant="thumbnail" alt="thumbnail image" />);
-  expect(wrapper.prop('className')).toContain('terra-Image--thumbnail');
   expect(wrapper.prop('alt')).toEqual('thumbnail image');
 });
 
 /* Image with fluid behavior */
 it('should render a thumbail image component with fluid behavior', () => {
-  const wrapper = shallow(<Image src="" variant="thumbnail" alt="thumbnail image" isFluid />);
-  expect(wrapper.prop('className')).toContain('terra-Image--thumbnail terra-Image--fluid');
+  const wrapper = render(<Image src="" variant="thumbnail" alt="thumbnail image" isFluid />);
+  expect(wrapper).toMatchSnapshot();
 });
 
 /* Image without fluid behavior */
 it('should render a thumbail image component with non-fluid behavior', () => {
-  const wrapper = shallow(<Image src="" variant="thumbnail" alt="thumbnail image" />);
-  expect(wrapper.prop('className')).toContain('terra-Image--thumbnail');
+  const wrapper = render(<Image src="" variant="thumbnail" alt="thumbnail image" />);
+  expect(wrapper).toMatchSnapshot();
 });
 
 /* Image with fluid behavior having a parent container */