+++ packages/terra-form/CHANGELOG.md
@@ -3,12 +3,15 @@ Changelog
 
 Unreleased
 ----------
+
+1.2.0 - (August 1, 2017)
+------------------
 ### Changed
+* Converted package to use CSS Modules
 * Update Field documentation
 * Change NumberField defaultValue propType to accept string or numeric values
 * Removed red ring for required fields.
 * Select and SelectField also accept prop 'options', a list of objects to set their value and display
-
 
 1.1.0 - (July 18, 2017)
 ------------------