+++ packages/terra-list/CHANGELOG.md
@@ -3,6 +3,9 @@ Changelog
 
 Unreleased
 ----------
+
+1.3.0 - (August 1, 2017)
+------------------
 ### Changed
 * Converted component to use CSS modules