+++ packages/terra-popup/CHANGELOG.md
@@ -3,8 +3,19 @@ Changelog
 
 Unreleased
 ----------
+
+1.2.0 - (August 1, 2017)
+------------------
+### Changed
+* Minor version bump
+
+1.1.0 - (August 1, 2017)
+------------------
 ### Added
 * Added requestFocus and releaseFocus props to support requesting and releasing focus.
+
+### Fixed
+* Prevent Invalid Props from breaking size logic. (Issue #679)
 
 1.0.0 - (July 26, 2017)
 ------------------